#77 rename : calculate -> calculations
