+++ fe/src/utils/calculations.ts
